@@ -1,21 +1,27 @@
 name: worker_service
-version: 0.5.2+3
+version: 0.7.4
 homepage: https://github.com/SunnyApp/worker_service
 description: A flutter package that works for native/web, and allows the creation of workers that offload work into a potentially parallel space
 environment: 
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies: 
   async: ^2.9.0
+  collection: ^1.16.0
   equatable: ^2.0.5
-  isolate: ^2.0.3
   js: ^0.6.4
   logging: ^1.0.2
   meta: ^1.8.0
-  service_worker: ^0.2.4
+  uuid: ^3.0.6
+  isolates: 
+    path: /Users/ericm/sunny/local_plugins/isolate
   logging_config: 
-    path: /Users/ericm/reliveit/checkout/SunnyApp/logging_config
-  sunny_dart: 
-    path: /Users/ericm/reliveit/checkout/SunnyApp/sunny_dart
+    path: /Users/ericm/sunny/local_plugins/logging_config
+  dartxx: 
+    path: /Users/ericm/sunny/local_plugins/dartxx
+dev_dependencies: 
+  pedantic: ^1.11.1
+  flutter_test: 
+    sdk: flutter
 flutter: 
   assets: 
     - bootstrap_dev.js
