@@ -1,35 +1,34 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
-import 'package:sunny_dart/helpers.dart';
 
+import '../utils.dart';
 import 'grunt_channel.dart';
 import 'grunt_registry.dart';
 import 'message.dart';
 import 'work.dart';
 
-WorkPhase workPhaseOf(int i) {
-  return WorkPhase.values
-      .firstWhere((element) => element.ordinal == i, orElse: () => null);
+WorkPhase? workPhaseOf(int? i) {
+  return WorkPhase.values.firstWhereOrNull((element) => element.ordinal == i);
 }
 
 class ErrorStack {
-  final Object error;
-  final StackTrace stack;
+  final Object? error;
+  final StackTrace? stack;
 
   ErrorStack(this.error, [this.stack]);
 
   const ErrorStack.of({
-    @required this.error,
-    @required this.stack,
+    required this.error,
+    required this.stack,
   });
 
-  factory ErrorStack.fromJson(map) {
+  static ErrorStack? fromJson(map) {
     if (map == null) return null;
     return ErrorStack.of(
       error: map['error']?.toString(),
-      stack: StackTrace.fromString(map['stack']?.toString()),
+      stack: StackTrace.fromString(map['stack']?.toString() ?? ''),
     );
   }
 
@@ -95,13 +94,13 @@ extension GruntExt on Grunt {
 mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
   PayloadHandler get fallbackEncoding => PayloadHandler.defaults;
 
-  PayloadHandler _encoding;
+  PayloadHandler? _encoding;
 
-  GruntChannel _channel;
-  Logger _log;
+  GruntChannel? _channel;
+  Logger? _log;
 
   @override
-  Logger get log => _log ??= Logger("$runtimeType");
+  Logger get log => _log ??= Logger("<< grunt:$key");
 
   @override
   final String jobId = uuid();
@@ -109,14 +108,14 @@ mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
   @override
   WorkPhase workPhase = WorkPhase.ready;
 
-  double total;
+  double? total;
   double progress = 0.0;
 
   dynamic params;
-  ErrorStack error;
+  ErrorStack? error;
 
-  String message;
-  Map<String, dynamic> state;
+  String? message;
+  Map<String, dynamic>? state;
 
   @override
   void initialize(GruntChannel channel, [params]) {
@@ -131,7 +130,7 @@ mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
 
   Future execute(dynamic params);
 
-  FutureOr onError(ErrorStack err) {}
+  FutureOr onError(ErrorStack? err) {}
 
   FutureOr onStop() {}
 
@@ -151,7 +150,8 @@ mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
     return next(payload);
   }
 
-  dynamic decodePayload(int contentType, dynamic content, PayloadDecoder next) {
+  dynamic decodePayload(
+      int? contentType, dynamic content, PayloadDecoder next) {
     return next(contentType, content);
   }
 
@@ -181,7 +181,7 @@ mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
   }
 
   void sendUpdate(
-      {String message, double progress, Map<String, dynamic> state}) {
+      {String? message, double? progress, Map<String, dynamic>? state}) {
     if (message != null) {
       this.message = message;
     }
@@ -219,6 +219,6 @@ mixin GruntMixin<SELF extends Grunt> implements Grunt, GruntFactory<SELF> {
   @override
   GruntChannel get channel =>
       _channel ??
-      illegalState(
-          "Attempting to communicate before channel has been established");
+      (throw Exception(
+          "Attempting to communicate before channel has been established"));
 }
