@@ -6,8 +6,7 @@ import 'dart:async';
 import 'dart:html';
 
 import 'package:js/js.dart';
-import 'package:sunny_dart/helpers/logging_mixin.dart';
-import 'package:meta/meta.dart';
+import 'package:logging/logging.dart';
 import 'package:worker_service/work/grunt_channel.dart';
 
 import '../message.dart';
@@ -26,24 +25,25 @@ abstract class XMLHttpRequestProgressEvent implements ProgressEvent {
   }
 }
 
-class WWDuplexChannel extends DuplexChannel with LoggingMixin {
+class WWDuplexChannel extends DuplexChannel {
+  final Logger log;
   final _inbound = StreamController<DecodedMessage>.broadcast();
   @override
   final PayloadHandler encoding;
 
-  WWDuplexChannel(this.encoding) {
-    print("Inside web worker, setting up onmessage listener for supervisor");
+  WWDuplexChannel(this.encoding, {String? channelName})
+      : log = Logger(channelName ?? 'webworker') {
+    log.info("Inside web worker, setting up onmessage listener for supervisor");
     onMessage = allowInterop((event) {
       try {
         if (event is MessageEvent) {
-          print("Message into ww ${event?.data}");
+          log.info("WebWorker received raw message: ${event.data}");
           _inbound.add(DecodedMessage.decoded(event.data, encoding));
-          print("Send message to stream...");
         } else {
-          print("Got non-message event: $event");
+          log.warning("Got non-message event: $event");
         }
-      } catch (e) {
-        print(e);
+      } catch (e, stack) {
+        log.severe("ERROR LISTENING TO EVENT: $e", e, stack);
       }
     });
   }
@@ -57,11 +57,12 @@ class WWDuplexChannel extends DuplexChannel with LoggingMixin {
   Stream<DecodedMessage> get inbound => _inbound.stream;
 
   @override
-  void send(int type, [dynamic payload, int contentType]) {
+  void send(int type, [dynamic payload, int? contentType]) {
     try {
-      log.info("Sending message $type with payload $payload to the other side");
+      log.info(
+          "Sending message ${MessageCode.get(type)} with payload $payload to the outer control");
       final _payload = encoding.encode(Payload(contentType, payload));
-      print("Message: $type with payload $payload");
+
       postMessage([
         type,
         _payload.header,
@@ -74,7 +75,9 @@ class WWDuplexChannel extends DuplexChannel with LoggingMixin {
   }
 }
 
-class WebDuplexChannel with LoggingMixin implements DuplexChannel {
+class WebDuplexChannel implements DuplexChannel {
+  final Logger log;
+
   /// The actual WebWorker instance
   final Worker worker;
 
@@ -83,21 +86,27 @@ class WebDuplexChannel with LoggingMixin implements DuplexChannel {
 
   final String debugLabel;
 
-  @override
-  String get loggerName => debugLabel;
-
   final _inbound = StreamController<DecodedMessage>.broadcast();
 
-  WebDuplexChannel(this.worker, this.encoding, {@required this.debugLabel}) {
+  WebDuplexChannel(this.worker, this.encoding, {required this.debugLabel})
+      : log = Logger(debugLabel) {
     worker.onMessage.forEach((element) {
       _inbound.add(DecodedMessage.decoded(element.data, encoding));
     });
   }
 
   @override
-  void send(int type, [dynamic payload, int contentType]) {
-    log.info("Sending message $type with payload $payload to the other side");
+  void send(int type, [dynamic payload, int? contentType]) {
     final _payload = encoding.encode(Payload(contentType, payload));
+    if (payload != null) {
+      log.info(
+          "Sending message $type with original:${payload.runtimeType}: encoded as header=${_payload.header} data=${_payload.data}");
+      log.info(
+          "  > encoded as Sending message $type with payload ${payload.runtimeType}");
+    } else {
+      log.info("Sending message $type with null payload ${_payload.header}");
+    }
+
     worker.postMessage([
       type,
       _payload.header,
