@@ -1,7 +1,9 @@
 // ignore_for_file: camel_case_types
 
 class grunt {
-  const grunt();
+  const grunt({this.logLevel = "INFO", this.logOutput = "console"});
+  final String logLevel;
+  final String logOutput;
 }
 
 class gruntMain {
