@@ -1,6 +1,8 @@
 import 'dart:async';
 
-import 'package:isolate/runner.dart';
+import 'package:isolates/runner.dart';
+import 'package:isolates/runner_factory.dart';
+import 'package:logging_config/logging_environment.dart';
 import 'common.dart';
 
 WorkerServicePlatform workerService() =>
@@ -8,3 +10,5 @@ WorkerServicePlatform workerService() =>
 
 /// Spawns a new isolate with any extra processing.  It's required to have a top-level spawner method
 Future<Runner> spawnRunner(RunnerBuilder factory) => throw "Not implemented";
+
+LoggingEnvironment workerLogEnvironment() => throw "Not implemented";
